### 2024-02-14 #6238 @thegargiulian
Hi @samhforbes, thanks for the invitation, but I won't be available to review for JOSS for the next two months
