### 2024-04-02 #6238 @samhforbes
Hi @rmhogervorst, @russellpierce would you be interested in reviewing this software for us? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
