### 2024-04-11 #6238 @samhforbes
@misken @RhysPeploe @cansavvy would you be interested in reviewing this software for us? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
