### 2024-04-11 #6238 @cansavvy
I'm definitely interested in being a reviewer for JOSS but have not worked with social science data so I might not be the best pick for this one.
