### 2024-04-12 #6238 @misken
I'm willing to do it with the following caveat. I'm also not a social science person and don't work with survey data (have never used Stata - I use R quite a bit).
