### 2024-04-12 #6238 @samhforbes
Hi @misken if you'd be comfortable I'd quite like your perspective on it as an R package. I'll have additional eyes on it myself as a social scientist and R user.
