### 2024-04-12 #6238 @misken
Yep, @samhforbes, I'm happy to review it. This will be my first JOSS review.
