### 2024-04-12 #6610 @editorialbot
[REVIEW]: jtools: Analysis and Presentation of Social Scientific Data
**Submitting author:** @jacob-long (<a href="http://orcid.org/0000-0002-1582-6214">Jacob A. Long</a>)
**Repository:** https://github.com/jacob-long/jtools
**Branch with paper.md** (empty if default branch): joss
**Version:** 2.3.0
