### 2024-04-12 jacob-long/jtools#149 @misken
Dependencies question for JOSS review
Hi @jacob-long, I'm one of the reviewers for your recently submitted JOSS paper. I'll be opening issues here regarding the review - which is at https://github.com/openjournals/joss-reviews/issues/6610.

I installed `jtools` in an R virtual env and was just reproducing the examples from the GitHub readme page. That page does mention that various `jtools` functionality depends on `sandwich` and `huxtable`. I attempted running the code before installing these and got appropriate error messages. After installing `huxtable` and rerunning the example using `export_summs`, then I got an error message saying I needed to install `brooms`. After installing `brooms`, that example worked. Further down, when trying `plot_summs`, I got the correct output but got errors in the console complaining about needing `brooms.mixed`. I see that these are all suggested packages, so I'm not saying these are bugs or problems. Just pointing out that some R newbies might get hung up on stuff like this. I'm not an R package dev so not really sure what the best practice is for handling this - maybe you're already doing what is suggested by CRAN.
