### 2024-04-12 #6610 @yjunechoe
I've been test-driving {jtools} for a while since I agreed to the review, so here's an early review. Overall, this is a well-designed, feature-rich package, and it was a pleasure to review. It offers a substantial contribution to data analysis and statistical modelling for research for use in the social (and other) sciences.

## Checklist comments

### Automated tests

I've tested most examples and tests and they run fine on my local machine, but I see some GHA workflows failing on the package repo - please give the failing actions another run and resolve any issues if they persist.

### Paper: statement of need/ state of the field

Since the package touches various areas of statistical and data analysis, the package's features overlap with many packages that are more specific in nature. As mentioned in the paper, these include packages like {marginaleffects}, {modelsummary}, {sjPlot}, etc. - I would also like to add to this list {ggeffects} and {gtsummary} (in so far as {jtools} provide reporting capability), and other members of the {easystats} ecosystem (such as {performance} and {parameters}).

I think it'd be helpful for the paper to not only mention the similarities, but also highlight the advantages and disadvantages to using {jtools} as opposed to using a combination of specialized packages. For example, the claim that "... {jtools} aims to improve the user interface and support more model types in some cases" could use some specific examples. Ideally, something like a table of comparisons (similar packages as columns, features as rows) would be extremely helpful for new users deciding whether {jtools} is right for them.

## Comments on software

### Hidden implicit dependencies

The package has several "soft" dependencies - these are listed in Suggests and only loaded when needed, as they should be. However, it seems that these soft dependencies are causing some confusion for users who do not have them installed (as in issues 132, 107). One thing that could help here is to use `rlang::check_installed()`, which would not only check a package's availability but also prompt the user to install the missing package to proceed. This could replace/standardize the various in-house checking logics like the below:

```r
  if (!requireNamespace("lme4", quietly = TRUE)) {
    stop("Package `lme4` needed for this function to work. Please install it.",
         call. = FALSE)
  }
```

Additionally, in at least one case, a package in Suggests is versioned but there is no version compatibility check at the point when it's loaded (`huxtable >= 3.0.0`). Dependencies listed in Suggests have to be manually checked for their version on load (this is again easy to do with `rlang::check_installed()`).

Lastly, speaking of dependencies and namespaces, it looks like this piece of code in `plot_coefs.R` is waiting for a refactoring, now that :

```r
    # Major kludge for methods clash between broom and broom.mixed
    # Making namespace environment with broom.mixed before generics
    # to try to put those methods in the search path
    # Will drop after update to broom 0.7.0
    if (requireNamespace("broom.mixed")) {
      nse <- as.environment(unlist(sapply(c(asNamespace("broom.mixed"),
                                            asNamespace("generics")),
                                          as.list)))
    } else {
      nse <- asNamespace("generics")
    }
```

### A tension in dependency versioning

This package has a long history (I count ~8years since the first commit), and I really applaud the effort to preserve some of the older behaviors and features for maximal reproducibility. However I also see some tension between the desire to integrate cutting-edge tools and the desire to maintain backwards compatibility for users on older R / R package versions.

This is most apparent in the versioning of packages in `Imports`. It lists `ggplot2 (>= 3.4.0)` (younger, released 2022) and `rlang (>= 0.3.0)` (older, released 2018). Typically, this kind of a division is appropriate - low-level, "back-end" dependencies to be as stable as possible, but that constraint should be relaxed for "front-end" dependencies, as updates to those packages usually bring about user-visible improvements. However, this overlooks the fact that `ggplot2 (>=3.4.0)` itself depends on the more recent `rlang (>=1.0.0)` - this ends up circumventing a lot of backwards-compatibility measures carefuly implemented in the package and maintained over the  in the package. For example, this `if` block is dead code because the condition will *never* trigger:

```r
    # Make sure rlang knows about "x" and "i" bullets
    if (utils::packageVersion("rlang") >= "0.4.2") {
```

I strongly urge the author to revisit these backwards compatibility workarounds scattered across the source code and, where appropriate, remove dead code. If possible, I further suggest that the author takes the JOSS review process as an opportunity to push an update to the package bringing it more up to date with the current state of the tidyverse ({rlang} bumped to the "production" release v1.0, {cli} used in place of {crayon}, etc.), and unapologically dropping some support for older behaviors that are not critical to the correctness of output.

### Outdated non-standard evaluation patterns

This is a minor comment, but inside `aes()`, `!!sym("col")` can now be `.data$col` and `!!col` can now be `.data[[col]]`. This brings {ggplot2} NSE patterns up to date with the most recent changes to tidyevaluation semantics in {rlang}, namely, the introduction of the `.data` and `.env` pronouns.

I think there are some other cases of NSE that could be simplified/streamlined - I check whether these can be updated in accordance with rlang v1.0 (again, I strongly encourage this version bump given what I said above)

---

### Misc.

I'll add more minor comments here as I they come up, but this summarizes most of my thoughts.
