### 2024-04-19 jacob-long/jtools#149 @gaborio
I was actually going to open a discussion about this, I use the package in my class and I forgot that in my machine I have those dependencies installed, but my students don't so they encountered errors. What is the logic for making them suggestions instead of dependencies?
