### 2024-04-20 jacob-long/jtools#149 @jacob-long
Thanks for raising this. These are, to some extent, decisions made long ago and not often re-examined so it's helpful to hear when/if they cause trouble for somebody. The initial reasoning was fairly straightforward in that I hoped to avoid "too many" hard dependencies and the functionality these were providing was probably not as core to the package as they may seem now.

That being said, my reaction to the issue is that probably all of these packages should be imports rather than suggests, with the possible exception of `huxtable`. I certainly aim to make this package particularly accessible to people who are relatively unfamiliar with R/programming and package management isn't going to be second nature to those folks — and I'm already making some other design choices that make things a little inefficient in the service of that audience so this would be consistent with that approach.

I will take a look at this soon and (most likely) make some changes.
