### 2024-04-29 #6610 @yjunechoe
On the note of tests, I can confirm the correctness of tests and they have good coverage, though they pass/fail to varying degrees due to the technicalities of how tests are implemented. Namely, it seems that the tests for plotting features (in `test-effect-plots.R` and `test-export-summs.R`) use **snapshot tests** on exactness of the graphical output (svg). I'll just note that this is extremely brittle to changes in the internals of upstream packages, namely `{grid}` and `{ggplot2}`. For example, all tests on plots would fail if `{ggplot2}` someday decides to draw axes right-to-left vs. left-to-right. Indeed, all plots currently fail the snapshot tests for such trivial reasons (so please update the snaps). There are of course pros and cons to this setup - it makes it trivial to write tests but at the cost of maintenance (as you're already aware, snaps require manual updates every time they get outdated). In case test longevity is a concern, I just wanted to point to another possibility for testing plots, which is to test specific properties read off of the ggplot2 object (e.g., via `layer_data()`).

Echoing @misken's comment, I think this also wraps up my thoughts for now. I hope that the suggestions given so far point to clear targets that are straightforward to address.
