### 2024-05-24 ba71ef3
fix facet_wrap title box removing upper box line
