### 2024-06-18 #6610 @samhforbes
Hi @jacob-long sorry I totally missed this last question.
If you are satisfied that you've addressed questions and concerns you can say so here (or in issues directly where relevant) and you can tag in the relevant reviewer where appropriate. From what I can see above none of these are "blocking" issues for the reviewers, so it's a question of addressing these in a way you feel improved your package!
