+++ pkgdown.yml
@@ -1,9 +1,8 @@
-pandoc: 2.19.2
-pkgdown: 2.0.7
+pandoc: 3.1.11
+pkgdown: 2.1.0
 pkgdown_sha: ~
 articles:
   effect_plot: effect_plot.html
   summ: summ.html
   svycor: svycor.html
-last_built: 2023-07-06T19:38Z
-
+last_built: 2024-07-31T12:49Z
