+++ pkgdown.yml
@@ -5,4 +5,4 @@ articles:
   effect_plot: effect_plot.html
   summ: summ.html
   svycor: svycor.html
-last_built: 2024-07-31T12:49Z
+last_built: 2024-07-31T16:02Z
