### 2024-08-01 39866e3
Switch to `cli` dependency from `crayon`
