+++ R/summ.R
@@ -20,7 +20,8 @@
 #'
 #' @export
 #' @importFrom stats nobs
-#'
+#' @import cli
+#' 
 
 
 summ <- function(model, ...) {
@@ -413,11 +414,11 @@ print.summ.lm <- function(x, ...) {
   }
 
   if (x$model.fit == TRUE && !is.null(x$modpval)) {
-    stats <- paste(italic("F"), "(", x$fnum, ",", x$fden, ") = ",
-        num_print(x$fstat, digits = x$digits), ", ", italic("p"), " = ",
+    stats <- paste(style_italic("F"), "(", x$fnum, ",", x$fden, ") = ",
+        num_print(x$fstat, digits = x$digits), ", ", style_italic("p"), " = ",
         num_print(x$modpval, digits = x$digits), "\n",
-        italic("R\u00B2 = "), num_print(x$rsq, digits = x$digits), "\n",
-        italic("Adj. R\u00B2 = "), num_print(x$arsq, digits = x$digits),
+        style_italic("R\u00B2 = "), num_print(x$rsq, digits = x$digits), "\n",
+        style_italic("Adj. R\u00B2 = "), num_print(x$arsq, digits = x$digits),
         sep = "")
     print_mod_fit(stats)
   }
@@ -851,9 +852,9 @@ print.summ.glm <- function(x, ...) {
       type <- "Linear regression"
     } else {
       type <- paste("Generalized linear model\n ",
-                    italic("Family:"),
+                    style_italic("Family:"),
                     as.character(x$lmFamily[1]), "\n ",
-                    italic("Link function:"),
+                    style_italic("Link function:"),
                     as.character(x$lmFamily[2]), sep = " ")
     }
     print_mod_info(missing = x$missing, n = x$n, dv = x$dv, type = type)
@@ -862,13 +863,13 @@ print.summ.glm <- function(x, ...) {
   if (x$model.fit == TRUE) {
     stats <- paste("\u03C7\u00B2(",
                   x$chisq$df,  ") = ", num_print(x$chisq$chi, x$digits), ", ",
-                  italic("p"), " = ", num_print(x$chisq$p, x$digits), "\n",
-                   italic("Pseudo-R\u00B2 (Cragg-Uhler)"), " = ",
+                  style_italic("p"), " = ", num_print(x$chisq$p, x$digits), "\n",
+                   style_italic("Pseudo-R\u00B2 (Cragg-Uhler)"), " = ",
                    num_print(x$rsq, digits = x$digits), "\n",
-                   italic("Pseudo-R\u00B2 (McFadden)"), " = ",
+                   style_italic("Pseudo-R\u00B2 (McFadden)"), " = ",
                    num_print(x$rsqmc, digits = x$digits), "\n",
-                   italic("AIC"), " = ", num_print(x$aic, x$digits),
-                   ", ", italic("BIC"), " = ", num_print(x$bic, x$digits),
+                   style_italic("AIC"), " = ", num_print(x$aic, x$digits),
+                   ", ", style_italic("BIC"), " = ", num_print(x$bic, x$digits),
                    sep = "")
     print_mod_fit(stats)
   }
@@ -1340,8 +1341,8 @@ print.summ.svyglm <- function(x, ...) {
     } else {
       # Otherwise just treat it like glm
       type <- paste("Analysis of complex survey design", "\n",
-                  italic("Family:"), as.character(x$lmFamily[1]),
-                  "\n", italic("Link function:"), as.character(x$lmFamily[2]),
+                  style_italic("Family:"), as.character(x$lmFamily[1]),
+                  "\n", style_italic("Link function:"), as.character(x$lmFamily[2]),
                   sep = " ")
     }
[…]
