+++ NAMESPACE
@@ -206,6 +206,7 @@ export(weights_tests)
 export(wgttest)
 export(wrap_str)
 export(wtd.sd)
+import(cli)
 import(ggplot2)
 import(rlang)
 importFrom(crayon,cyan)
