+++ R/summ_helpers.R
@@ -86,7 +86,7 @@ round_df_char <- function(df, digits, pad = " ", na_vals = NA) {
 #' @param type One of `"HC3"`, `"const"`, `"HC"`, `"HC0"`, `"HC1"`,
 #'  `"HC2"`, `"HC4"`, `"HC4m"`, `"HC5"`. See [sandwich::vcovHC()] for some 
 #'  more details on these choices. Note that some of these do not work for
-#'  clustered standard errors (see sandwich::vcovCL()]). 
+#'  clustered standard errors (see [sandwich::vcovCL()]). 
 #' @param data The data used to fit the model. Default is to just get the 
 #'  `model.frame` from `model`. 
 #' @param cluster If you want clustered standard errors, either a string naming
@@ -112,11 +112,6 @@ round_df_char <- function(df, digits, pad = " ", na_vals = NA) {
 #' @rdname get_robust_se
 get_robust_se <- function(model, type = "HC3", cluster = NULL, 
                           data = model.frame(model), vcov = NULL) {
-  
-  if (!requireNamespace("sandwich", quietly = TRUE)) {
-    stop_wrap("When using robust SEs you need to have the \'sandwich\' 
-              package.", call. = FALSE)
-  }
   
   if (type == TRUE) {
     type <- "HC3"
