+++ NAMESPACE
@@ -209,11 +209,6 @@ export(wtd.sd)
 import(cli)
 import(ggplot2)
 import(rlang)
-importFrom(crayon,cyan)
-importFrom(crayon,inverse)
-importFrom(crayon,italic)
-importFrom(crayon,red)
-importFrom(crayon,underline)
 importFrom(generics,glance)
 importFrom(generics,tidy)
 importFrom(ggplot2,aes_string)
