### 2024-08-02 e80f341
Use `rlang` to check for missing packages
