+++ man/gscale.Rd
@@ -164,8 +164,8 @@ regression models. On request, it will center and/or standardize variables
 before printing its output.
 
 Other standardization: 
-\code{\link{center_mod}()},
 \code{\link{center}()},
+\code{\link{center_mod}()},
 \code{\link{scale_mod}()},
 \code{\link{standardize}()}
 }
