+++ NAMESPACE
@@ -224,6 +224,8 @@ importFrom(grDevices,rgb)
 importFrom(magrittr,"%<>%")
 importFrom(magrittr,"%>%")
 importFrom(pander,pandoc.table.return)
+importFrom(rlang,check_installed)
+importFrom(rlang,is_installed)
 importFrom(stats,"contrasts<-")
 importFrom(stats,AIC)
 importFrom(stats,BIC)
