+++ pkgdown.yml
@@ -5,4 +5,4 @@ articles:
   effect_plot: effect_plot.html
   summ: summ.html
   svycor: svycor.html
-last_built: 2024-08-02T01:53Z
+last_built: 2024-08-02T14:37Z
