### 2024-08-02 3df7072
Eliminate old R 3.5 compatibility namespace kludges
