+++ R/export_summ.R
@@ -355,12 +355,7 @@ export_summs <- function(...,
 }
 
 #' @rdname glance.summ
-#' @rawNamespace 
-#' if (getRversion() >= "3.6.0") {
-#'  S3method(generics::tidy, summ)
-#' } else {
-#'  export(tidy.summ)
-#' }
+#' @export
 
 tidy.summ <- function(x, conf.int = FALSE, conf.level = .95, ...) {
 
@@ -471,12 +466,7 @@ tidy.summ <- function(x, conf.int = FALSE, conf.level = .95, ...) {
 }
 
 #' @rdname glance.summ
-#' @rawNamespace 
-#' if (getRversion() >= "3.6.0") {
-#'  S3method(generics::tidy, summ.merMod)
-#' } else {
-#'  export(tidy.summ.merMod)
-#' }
+#' @export
 
 tidy.summ.merMod <- function(x, conf.int = FALSE, conf.level = .95, ...) {
 
@@ -580,31 +570,7 @@ tidy.summ.merMod <- function(x, conf.int = FALSE, conf.level = .95, ...) {
 #'  \code{\link[generics]{glance}}
 #'
 #' @rdname glance.summ
-#' @rawNamespace 
-#' if (getRversion() >= "3.6.0") {
-#'  S3method(generics::glance, summ.lm)
-#' } else {
-#'  export(glance.summ.lm)
-#' }
-#' 
-#' if (getRversion() >= "3.6.0") {
-#'  S3method(generics::glance, summ.glm)
-#' } else {
-#'  export(glance.summ.glm)
-#' }
-#' 
-#' if (getRversion() >= "3.6.0") {
-#'  S3method(generics::glance, summ.svyglm)
-#' } else {
-#'  export(glance.summ.svyglm)
-#' }
-#' 
-#' if (getRversion() >= "3.6.0") {
-#'  S3method(generics::glance, summ.merMod)
-#' } else {
-#'  export(glance.summ.merMod)
-#' }
-
+#' @export
 
 glance.summ.lm <- function(x, ...) {
 
