+++ R/internal.R
@@ -348,6 +348,7 @@ test_coefs.default <- function(x, the_vcov = NULL, df = NULL, ...) {
 
 }
 
+#' @exportS3Method NULL
 test_coefs.glm <- function(x, the_vcov = NULL, df = Inf, ...) {
   # Only difference is default DF
   test_coefs.default(x, the_vcov = the_vcov, df = df, ...)
@@ -419,12 +420,7 @@ predict_rob <- function(model, .vcov = vcov(model), newdata = NULL,
 
 #' @importFrom tibble tibble as_tibble
 #' @importFrom stats confint
-#' @rawNamespace 
-#' if (getRversion() >= "3.6.0") {
-#'   S3method(generics::tidy, summary.glht)
-#' } else {
-#'   export(tidy.summary.glht)
-#' }
+#' @export
 tidy.summary.glht <- function(x, conf.int = FALSE, conf.level = 0.95, ...) {
   lhs_rhs <- tibble(lhs = rownames(x$linfct), rhs = x$rhs)
   coef <- as_tibble(x$test[c("coefficients", "sigma", 
