+++ man/gscale.Rd
@@ -163,7 +163,7 @@ open access)
 regression models. On request, it will center and/or standardize variables
 before printing its output.
 
-Other standardization: 
+standardization, scaling, and centering tools
 \code{\link{center}()},
 \code{\link{center_mod}()},
 \code{\link{scale_mod}()},
