+++ man/standardize.Rd
@@ -53,7 +53,7 @@ standardize(mtcars, vars = "qsec")
 
 }
 \seealso{
-Other standardization: 
+standardization, scaling, and centering tools
 \code{\link{center}()},
 \code{\link{center_mod}()},
 \code{\link{gscale}()},
