+++ man/scale_mod.Rd
@@ -134,7 +134,7 @@ ed.). Mahwah, NJ: Lawrence Erlbaum Associates, Inc.
 \code{\link{interact_plot}} creates attractive, user-configurable plots of
 interaction models.
 
-Other standardization: 
+standardization, scaling, and centering tools
 \code{\link{center}()},
 \code{\link{center_mod}()},
 \code{\link{gscale}()},
