+++ man/center_mod.Rd
@@ -106,7 +106,7 @@ ed.). Mahwah, NJ: Lawrence Erlbaum Associates, Inc.
 \code{\link{interact_plot}} creates attractive, user-configurable plots of
 interaction models.
 
-Other standardization: 
+standardization, scaling, and centering tools
 \code{\link{center}()},
 \code{\link{gscale}()},
 \code{\link{scale_mod}()},
