+++ NAMESPACE
@@ -138,6 +138,8 @@ export(weights_tests)
 export(wgttest)
 export(wrap_str)
 export(wtd.sd)
+import(broom)
+import(broom.mixed)
 import(cli)
 import(ggplot2)
 import(rlang)
