### 2024-08-03 d7c01dd
Add dependency checking to GA
