+++ .github/workflows/recheck.yml
@@ -8,11 +8,11 @@ on:
           - strong
           - most
     
-  name: Reverse dependency check
-    
-  jobs:
-    revdep_check:
-      name: Reverse check ${{ inputs.which }} dependents
-      uses: r-devel/recheck/.github/workflows/recheck.yml@v1
-      with:
-        which: ${{ inputs.which }}
+name: Reverse dependency check
+
+jobs:
+  revdep_check:
+    name: Reverse check ${{ inputs.which }} dependents
+    uses: r-devel/recheck/.github/workflows/recheck.yml@v1
+    with:
+    which: ${{ inputs.which }}
