+++ .github/workflows/recheck.yml
@@ -15,4 +15,4 @@ jobs:
     name: Reverse check ${{ inputs.which }} dependents
     uses: r-devel/recheck/.github/workflows/recheck.yml@v1
     with:
-    which: ${{ inputs.which }}
+        which: ${{ inputs.which }}
