### 2024-08-04 747c6c7
Add facet.by to `effect_plot()`
Closes #147
