### 2024-08-04 ac3fae7
Support `fixest_multi` objects in `plot_coefs()`
Closes #123
