+++ R/plot_coefs.R
@@ -64,6 +64,13 @@
 #'   dependent variable, that can be specified here. If NULL, it is assumed you 
 #'   want coefficients for the location/mean parameter, not the distributional
 #'   parameter(s).
+#' @param coefs.match This modifies the way the `coefs` and `omit.coefs` 
+#'   arguments are interpreted. The default `"exact"` which represents the 
+#'   legacy behavior, will include/exclude coefficients that match exactly
+#'   with your inputs to those functions. If `"regex"`, `coefs` and
+#'   `omit.coefs` are used as the `pattern` argument for [grepl()] matching
+#'   the coefficient names. Note that using `"regex"` means you will be unable
+#'   to override the default coefficient names via a named vector.
 #' @return A ggplot object.
 #' @details A note on the distinction between `plot_summs` and `plot_coefs`:
 #'   `plot_summs` only accepts models supported by [summ()] and allows users
@@ -120,7 +127,8 @@ plot_summs <- function(..., ci_level = .95, model.names = NULL, coefs = NULL,
                        line.size = c(0.8, 2), legend.title = "Model",
                        groups = NULL, facet.rows = NULL, facet.cols = NULL,
                        facet.label.pos = "top", color.class = colors, 
-                       resp = NULL, dpar = NULL) {
+                       resp = NULL, dpar = NULL, 
+                       coefs.match = c("exact", "regex")) {
   
   # Capture arguments
   dots <- list(...)
@@ -151,7 +159,7 @@ plot_summs <- function(..., ci_level = .95, model.names = NULL, coefs = NULL,
                     groups = list(groups), facet.rows = facet.rows,
                     facet.cols = facet.cols, facet.label.pos = facet.label.pos, 
                     color.class = color.class, resp = resp, dpar = dpar,
-                    ex_args))
+                    coefs.match = coefs.match, ex_args))
   
   do.call("plot_coefs", args = args)
   
@@ -171,7 +179,8 @@ plot_coefs <- function(..., ci_level = .95, inner_ci_level = NULL,
                        line.size = c(0.8, 2), legend.title = "Model", 
                        groups = NULL, facet.rows = NULL, facet.cols = NULL,
                        facet.label.pos = "top", color.class = colors,
-                       resp = NULL, dpar = NULL) {
+                       resp = NULL, dpar = NULL, 
+                       coefs.match = c("exact", "regex")) {
   
   if (!is.numeric(line.size[1])) stop_wrap("line.size must be a number (or two
     numbers in a vector).")
@@ -219,6 +228,8 @@ plot_coefs <- function(..., ci_level = .95, inner_ci_level = NULL,
     ex_args <- NULL
   }
   
+  coefs.match <- match.arg(coefs.match)
+
   if (!is.null(omit.coefs) && !is.null(coefs)) {
     if (any(omit.coefs %nin% c("(Intercept)", "Intercept"))) {
       msg_wrap("coefs argument overrides omit.coefs argument. Displaying
@@ -235,7 +246,8 @@ plot_coefs <- function(..., ci_level = .95, inner_ci_level = NULL,
   # Create tidy data frame combining each model's tidy output
   tidies <- make_tidies(mods = mods, ex_args = ex_args, ci_level = ci_level,
                         model.names = model.names, omit.coefs = omit.coefs,
-                        coefs = coefs, resp = resp, dpar = dpar)
+                        coefs = coefs, resp = resp, dpar = dpar, 
+                        coefs.match = coefs.match)
   
   n_models <- length(unique(tidies$model))
   
[…]
