### 2024-08-06 717422b
Fix empty weights variable name
Closes #156
