+++ paper.md
@@ -101,21 +101,14 @@ Finally, @spalti2023 use `weights_tests()` to
 assess the sensitivity of their estimates to the influence of survey sampling
 weights in their research on scientific misperceptions.
 
-+-------------------------------+--------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------+-----------------------------------------------------------------------+------------------------------------------------------------------------------------+---------------------------------------------------------+--------------------------------------------------+
-| **Feature**                      | `jtools`                                                                                              | `marginaleffects`                                                                                 | `modelsummary`                                                       | `sjPlot`                                                             | `ggeffects`                                                                       | `gtsummary`                                            | `parameters`                                        |
-+-------------------------------+--------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------+-----------------------------------------------------------------------+------------------------------------------------------------------------------------+---------------------------------------------------------+--------------------------------------------------+
-| **Model information**              | Flexible summaries, robust SEs, VIFs, centered and scaled variables, model fit statistics             | Marginal effects, contrasts, equivalence tests                                                    | All information available via `broom` and `parameters`, plus robust SEs | Flexible summaries, robust SEs, scaled variables, model fit statistics | Marginal effects, contrasts, equivalence tests (from `marginaleffects` or `emmeans`) | Information available via `broom`, pairwise comparisons  | Coefficient estimates, intervals, test statistics  |
-+-------------------------------+--------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------+-----------------------------------------------------------------------+------------------------------------------------------------------------------------+---------------------------------------------------------+--------------------------------------------------+
-| **Supported models**               | lm, glm, mixed models, survey data for summaries, many more for visualization                         | Over 100 types                                                                                    | All models supported by `broom` and `parameters`                      | lm, glm, mixed models, Bayesian models                                      | Large number of models                                                             | All that are supported by `broom`                       | Large number, including structural models          |
-+-------------------------------+--------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------+-----------------------------------------------------------------------+------------------------------------------------------------------------------------+---------------------------------------------------------+--------------------------------------------------+
-| **Visualization**                  | Effect plots, coefficient plots (`ggplot2`)                                                           | Visualization of marginal effects, similar to `ggeffects`                                         | Coefficient plots (`ggplot2`)                                         | Coefficient plots (`ggplot2`)                                        | Plots for effects, predictions, etc.                                                  | Experimental coefficient plot functions                  | Visualization of model parameters                 |
-+-------------------------------+--------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------+-----------------------------------------------------------------------+------------------------------------------------------------------------------------+---------------------------------------------------------+--------------------------------------------------+
-| **Model summary export options**   | Coefficient plots, tables to PDF/html/Word via `huxtable`                                             | Limited (console-focused)                                                                         | Most extensive, customizable, easy to use (Markdown, LaTeX, HTML, etc.) | To HTML                                                            | To HTML and Markdown                                                                | Extensive (Word, PDF, HTML, etc.)                        | To HTML and Markdown                              |
[…]
