### 2024-08-09 7b1e203
Remove table, add more written comparisons
