+++ paper.bib
@@ -172,4 +172,26 @@ Tests},
   annotation = {00199}
 }
 
+@Article{ggeffects,
+  title = {ggeffects: Tidy Data Frames of Marginal Effects from Regression Models.},
+  volume = {3},
+  doi = {10.21105/joss.00772},
+  number = {26},
+  journal = {Journal of Open Source Software},
+  author = {Daniel Lüdecke},
+  year = {2018},
+  pages = {772},
+}
+
+@Article{gtsummary,
+  author = {Daniel D. Sjoberg and Karissa Whiting and Michael Curry and Jessica A. Lavery and Joseph Larmarange},
+  title = {Reproducible Summary Tables with the gtsummary Package},
+  journal = {{The R Journal}},
+  year = {2021},
+  url = {https://doi.org/10.32614/RJ-2021-053},
+  doi = {10.32614/RJ-2021-053},
+  volume = {13},
+  issue = {1},
+  pages = {570-580},
+}
 
