### 2024-08-09 95fe794
Get rid of stale check for `rlang` version
