+++ R/zzz.R
@@ -22,9 +22,5 @@
 #' @importFrom utils packageVersion
 .onLoad <- function(libname, pkgname) {
   pkgconfig::set_config("tibble::rownames" = NA)
-  if (packageVersion("rlang") < "0.3.2") {
-    options("rlang__backtrace_on_error" = "branch")
-  } else {
-    options("rlang_backtrace_on_error" = "branch")
-  }
+  options("rlang_backtrace_on_error" = "branch")
 }
