### 2024-08-09 #6610 @jacob-long
Hi all, and thanks for the patience. I'm ready to report on some changes. Most of my efforts have been focused on the package rather than the paper although I have revised the paper as well.

* Outside of reviewer requests, I've addressed a number of bugs, requests, and so on that had been lingering in my repository.
* Both reviewers mentioned the use of dependencies. As a result of those discussions, I've made the extensively-used `broom`, `broom.mixed`, and `sandwich` packages required dependencies rather than "Suggests."
* I've done a pass to try to remove outdated compatibility fixes and stale version requirements.
* I've made a number of changes to the use of `rlang`, including the use of its tool to check for missing packages (usually) and for non-standard evaluation in the plotting functions. `cli` is now used in place of `crayon`.
* I investigated changing the automated tests for plots. For now, I've decided not to make any changes, although I do recognize the status quo isn't perfect. Simply testing the data doesn't quite assure me, although I may carefully introduce testing for the layer data for some tests in the future. I have opted for SVG output rather than PNG because, at least last time I checked, PNG outputs are not consistent across platforms and I occasionally develop on Windows. But I'm still thinking about testing strategies that will leave me comfortable that I catch any changes/regressions. And there are such a large number of these tests that I want to make sure that any change in strategy will be worth the substantial effort in rewriting them all and making sure I haven't lost any coverage in the process. The brittleness is a downside but also an upside since it has on one or two occasions alerted me to issues affecting the appearance of outputs that I thought were important even if not literally "breaking."
* I have added more discussion of the differences between `jtools` and some other packages, including additional ones raised by reviewers. I have to confess that I spent ~two workdays trying to make a legible table before giving up for reasons related to both formatting and content. I'm open to revisiting it but decided I've pulled enough of my hair out for now :D

I think that's all I wanted to say, for now! I believe I've regenerated the PDF above.
