### 2024-08-10 #6610 @yjunechoe
Thanks for all the effort you've put into this revision! I feel that the submission is in a great place and I have no further suggestions - I'm happy for `jtools` and the paper to be published in the current state.

(for clarity to the editor) I'll follow up on the points above to additionally note that:

- Tests: I felt obligated to comment on the snapshot tests for the checklist since it was failing, but it's a perfectly valid method of testing (and I see that they've been updated to pass the tests), so I don't need to see any changes here.
- Paper: My checklist comments are fully addressed by the paragraph of comparisons to `marginaleffects`, `ggeffects`, etc. I agree on the characterization of `jtools` as perhaps more user-friendly and immediately usable out of the box due in part to its many-arguments design, as opposed to the many-functions design in some of the related packages. This clarifies enough for me and I do not need to see a physical table of comparisons here.
