+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: jtools
 Type: Package
 Title: Analysis and Presentation of Social Scientific Data
-Version: 2.2.3.9999
+Version: 2.3.0
 Authors@R: person(c("Jacob","A."), "Long", email = "jacob.long@sc.edu",
   role = c("aut", "cre"), comment = c(ORCID = "0000-0002-1582-6214"))
 Description: This is a collection of tools for more efficiently understanding 
@@ -47,7 +47,7 @@ Enhances:
   brms,
   quantreg,
   rstanarm
-RoxygenNote: 7.3.2.9000
+RoxygenNote: 7.3.2
 VignetteBuilder: knitr
 Roxygen: list(markdown = TRUE)
 Depends: 
