+++ NAMESPACE
@@ -84,7 +84,6 @@ S3method(summ,rq)
 S3method(summ,svyglm)
 S3method(tidy,summ)
 S3method(tidy,summ.merMod)
-S3method(tidy,summary.glht)
 export("%just%")
 export("%just%<-")
 export("%nin%")
@@ -169,7 +168,6 @@ importFrom(stats,binomial)
 importFrom(stats,coef)
 importFrom(stats,coefficients)
 importFrom(stats,complete.cases)
-importFrom(stats,confint)
 importFrom(stats,contr.treatment)
 importFrom(stats,cooks.distance)
 importFrom(stats,cor)
@@ -217,7 +215,6 @@ importFrom(stats,vcov)
 importFrom(stats,weighted.mean)
 importFrom(stats,weights)
 importFrom(tibble,as_tibble)
-importFrom(tibble,tibble)
 importFrom(utils,.S3methods)
 importFrom(utils,getFromNamespace)
 importFrom(utils,getS3method)
