+++ man/theme_apa.Rd
@@ -75,9 +75,9 @@ Facetting is supported, but APA guidelines are considerably less clear for
 such situations.
 
 This theme was created with inspiration from Rudolf Cardinal's
-\href{http://egret.psychol.cam.ac.uk/statistics/R/graphs2.html}{code}, which
-required updating for newer versions of \code{ggplot2} and adaptations for
-APA style.
+\href{https://web.archive.org/web/20220616072522/http://egret.psychol.cam.ac.uk/statistics/R/graphs2.html}{code},
+which required updating for newer versions of \code{ggplot2} and
+adaptations for APA style.
 }
 \examples{
 # Create plot with ggplot2
