### 2024-08-21 #6610 @samhforbes
@jacob-long it looks like you have some CRAN DOIs that can be added for marginaleffects and sjPlot if you wouldn't mind.
If you could also archive this release and let me know the DOI and version number I can add this information, and proceed towards accept.
