### 2024-08-25 cb681ca
Add DOI for sjPlot and marginaleffects
