+++ paper.bib
@@ -14,6 +14,7 @@
   author = {Daniel Lüdecke},
   note = {R package version 2.8.15},
   url = {https://CRAN.R-project.org/package=sjPlot},
+  doi = {10.32614/CRAN.package.sjPlot},
 }
 
 @Article{see,
@@ -63,9 +64,10 @@
     title = {marginaleffects: Predictions, Comparisons, Slopes, Marginal Means, and Hypothesis
 Tests},
     author = {Vincent Arel-Bundock},
-    year = {2023},
-    note = {R package version 0.17.0},
+    year = {2024},
+    note = {R package version 0.21.0},
     url = {https://CRAN.R-project.org/package=marginaleffects},
+    doi = {10.32614/CRAN.package.marginaleffects},
 }
 
 @article{sutin2023,
