### 2024-08-25 #6610 @jacob-long
Roger that! Just pushed a commit with DOIs added for marginaleffects and sjPlot.

I just tagged the latest release on Github, version [2.3.0](https://github.com/jacob-long/jtools/releases/tag/2.3.0).

Zenodo DOI for archived release: [10.5281/zenodo.13371207](https://zenodo.org/doi/10.5281/zenodo.13371207)
