### 2024-09-02 #6610 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
